% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{alpha_to_half_life}
\alias{alpha_to_half_life}
\alias{half_life_to_alpha}
\title{Convert between OU adaptation rate and phylogenetic half-life}
\usage{
alpha_to_half_life(alpha)

half_life_to_alpha(half.life)
}
\arguments{
\item{alpha}{non-negative OU adaptation rate.}

\item{half.life}{positive phylogenetic half-life.}
}
\value{
A numeric vector on the converted scale.
}
\description{
Convert between OU adaptation rate and phylogenetic half-life
}
