% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{evolutionary_vcov}
\alias{evolutionary_vcov}
\title{Evolutionary innovation covariance from a fitted model}
\usage{
evolutionary_vcov(object)
}
\arguments{
\item{object}{fitted object of class \code{"l1ou"}.}
}
\value{
The evolutionary innovation covariance among traits.
}
\description{
Evolutionary innovation covariance from a fitted model
}
