% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.l1ou}
\alias{plot.l1ou}
\title{Visualizes a shift configuration: tree and trait(s)}
\usage{
\method{plot}{l1ou}(
  x,
  palette = NA,
  edge.shift.ann = TRUE,
  edge.shift.adj = c(0.5, -0.025),
  edge.label = c(),
  asterisk = TRUE,
  edge.label.ann = FALSE,
  edge.label.adj = c(0.5, 1),
  edge.label.pos = NA,
  edge.ann.cex = 1,
  plot.bar = TRUE,
  bar.axis = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class l1ou returned by \pkg{kfl1ou}.}

\item{palette}{vector of colors, of size the number of shifts plus one. The last element is the color for the background regime (regime at the root).}

\item{edge.shift.ann}{logical. If TRUE, annotates edges by shift values.}

\item{edge.shift.adj}{adjustment argument to give to edgelabel() for labeling edges by shift values.}

\item{edge.label}{vector of size number of edges.}

\item{asterisk}{logical. If TRUE, the shift positions will be annotated by "*". It is useful for gray scale plots.}

\item{edge.label.ann}{logical. If TRUE, annotates edges by labels in tree$edge.label, if non-empty, or edge.label.}

\item{edge.label.adj}{adjustment argument to give to edgelabel() for labeling edges.}

\item{edge.label.pos}{relative position of the edge.label on the edge. 0 for the beginning of the edge and 1 for the end of the edge.}

\item{edge.ann.cex}{amount by which the annotation text should be magnified relative to the default.}

\item{plot.bar}{logical. If TRUE, the bars corresponding to the trait values will be plotted.}

\item{bar.axis}{logical. If TRUE, the axis of of trait(s) range will be plotted.}

\item{...}{further arguments to be passed on to plot.phylo.}
}
\value{
none.
}
\description{
Plots the tree annotated to show the edges with a shift, and the associated trait data side by side.
Each trait is shown standardized, to visually highlight the species with low,
average or high values. In other words, the axis scale shows the values of
(trait - m)/sd where m is the observed trait mean and sd is the observed trait
standard deviation. The values on the left and right side of the axis are the
mininum and maximum of the standardized trait: (min-m)/sd and (max-m)/sd.
The place where the bars start corresponds to the mean of the original trait values.
}
\examples{

data(lizard.traits, lizard.tree)
keep <- lizard.tree$tip.label[1:15]
tree <- ape::drop.tip(lizard.tree, setdiff(lizard.tree$tip.label, keep))
tree <- ape::reorder.phylo(tree, "postorder")
Y <- lizard.traits[keep, 1]
eModel <- estimate_shift_configuration(tree, Y, criterion="AICc", max.nShifts=2)
nEdges <- ape::Nedge(tree)
ew <- rep(1,nEdges) 
if (length(eModel$shift.configuration) > 0) {
  ew[eModel$shift.configuration] <- 3
}
plot(eModel, cex=0.5, label.offset=0.02, edge.width=ew)

}
