% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{summary.l1ou}
\alias{summary.l1ou}
\title{Prints out a summary of the model}
\usage{
\method{summary}{l1ou}(object, nTop.scores = 5, ...)
}
\arguments{
\item{object}{object of class l1ou returned by \pkg{kfl1ou}.}

\item{nTop.scores}{number of top scores and shift configuration to print out.}

\item{...}{further arguments.}
}
\value{
none.
}
\description{
prints out a summary of the model
}
\examples{

data(lizard.traits, lizard.tree)
keep <- lizard.tree$tip.label[1:15]
tree <- ape::drop.tip(lizard.tree, setdiff(lizard.tree$tip.label, keep))
tree <- ape::reorder.phylo(tree, "postorder")
Y <- lizard.traits[keep, 1]
eModel <- estimate_shift_configuration(tree, Y, criterion="AICc", max.nShifts=2)
summary(eModel)

}
