% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfl1ou-package.R
\docType{package}
\name{kfl1ou-package}
\alias{kfl1ou}
\alias{kfl1ou-package}
\title{kfl1ou: OU shift detection and comparative trait modelling}
\description{
Internal package declarations used to generate the namespace.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kfuku52.github.io/kfl1ou/}
  \item \url{https://github.com/kfuku52/kfl1ou}
  \item Report bugs at \url{https://github.com/kfuku52/kfl1ou/issues}
}

}
\author{
\strong{Maintainer}: Kenji Fukushima \email{kenji.fukushima@nig.ac.jp}

Authors:
\itemize{
  \item Kenji Fukushima \email{kenji.fukushima@nig.ac.jp}
  \item Mohammad Khabbazian
}

Other contributors:
\itemize{
  \item Cecile Ane [contributor]
  \item Qing Yu Yu (Sabrina) [contributor]
}

}
\keyword{internal}
