% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{profile.l1ou}
\alias{profile.l1ou}
\title{Prints out a summary of the shift configurations investigated by \code{\link{estimate_shift_configuration}}}
\usage{
\method{profile}{l1ou}(fitted, ...)
}
\arguments{
\item{fitted}{object of class l1ou returned by \pkg{kfl1ou}.}

\item{...}{further arguments.}
}
\value{
\item{shift.configurations}{list of shift configurations sorted by number of shifts.}
\item{scores}{numeric vector of scores corresponding to shift.configurations.}
\item{nShifts}{integer vector giving the number of shifts in each configuration.}
}
\description{
prints the list of the shift configurations sorted by number of shifts and corresponding ic scores.
}
\examples{

data(lizard.traits, lizard.tree)
keep <- lizard.tree$tip.label[1:15]
tree <- drop.tip(lizard.tree, setdiff(lizard.tree$tip.label, keep))
tree <- reorder(tree, "postorder")
Y <- lizard.traits[keep, 1]
eModel <- estimate_shift_configuration(tree, Y, criterion="AICc", max.nShifts=2)
model.profile  <- profile(eModel)
plot(model.profile$nShifts, model.profile$scores)

}
